move eq-epsilon related stuff to a new file (TestCommonEqEpsilon.hs)
